changed type durata_erogazione_sec in integer

+++ src/data_prep/feature_selection.py
@@ -122,7 +122,7 @@ def colonna_durata_erogazione(df:pd.DataFrame) -> pd.DataFrame:
     df['ora_inizio_erogazione'] = pd.to_datetime(df['ora_inizio_erogazione'], utc=True, errors='coerce')
     df['ora_fine_erogazione'] = pd.to_datetime(df['ora_fine_erogazione'], utc=True, errors='coerce')
 
-    df['durata_erogazione_sec'] = (df['ora_fine_erogazione'] - df['ora_inizio_erogazione']).dt.total_seconds()
+    df['durata_erogazione_sec'] = (df['ora_fine_erogazione'] - df['ora_inizio_erogazione']).dt.total_seconds().astype(int)
     
     return df
